% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htm2txt.R
\name{displaytxt}
\alias{displaytxt}
\title{Display a text in a web page at a certain URL}
\usage{
displaytxt(URL)
}
\arguments{
\item{URL}{a string containing the URL of a web page}
}
\description{
Display a text in a web page at a certain URL
}
\examples{
displaytxt("https://cran.r-project.org/")
}
