% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htm2txt.R
\name{gettxt}
\alias{gettxt}
\title{Extract simple plain texts from a web page at a certain URL}
\usage{
gettxt(URL, encoding = "UTF-8", ...)
}
\arguments{
\item{URL}{A character indicating the URL of a web page.}

\item{encoding}{Encoding method (e.g., "UTF-8", "latin1", "bytes", "unknown", etc.).}

\item{...}{Other \code{\link{htm2txt}} arguments.}
}
\value{
A character containing plain texts converted from the htm document at the URL.
}
\description{
Extract simple plain texts from a web page at a certain URL
}
\examples{
text = gettxt("https://www.wikipedia.org/")
}
