% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htm2txt.R
\name{browse}
\alias{browse}
\title{Display a plain text in a web page at a certain URL}
\usage{
browse(URL, encoding = "UTF-8")
}
\arguments{
\item{URL}{a string containing the URL of a web page.}

\item{encoding}{a string indicating the encoding method (e.g., "UTF-8", "latin1", "bytes", "unknown", etc.).}
}
\value{
None (invisible NULL).
}
\description{
Display a plain text in a web page at a certain URL
}
\examples{
browse("https://CRAN.R-project.org/package=htm2txt")
}
