% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htm2txt.R
\name{browse}
\alias{browse}
\title{Display simple plain texts in a web page at a certain URL}
\usage{
browse(URL, ...)
}
\arguments{
\item{URL}{A character indicating the URL of a web page.}

\item{...}{Other \code{\link{gettxt}} arguments.}
}
\value{
None (invisible NULL).
}
\description{
Display simple plain texts in a web page at a certain URL
}
\examples{
browse("https://www.wikipedia.org/")
}
