% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htm2txt.R
\name{crawltxt}
\alias{crawltxt}
\title{Extract a text from a web page at a certain URL}
\usage{
crawltxt(URL)
}
\arguments{
\item{URL}{a string containing the URL of a web page}
}
\value{
a string containing texts converted from the html document at the URL.
}
\description{
Extract a text from a web page at a certain URL
}
\examples{
text = crawltxt("https://cran.r-project.org/")
}
