% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htm2txt.R
\name{htm2txt}
\alias{htm2txt}
\title{Wipe out tags in a html document and return a text in it}
\usage{
htm2txt(htm)
}
\arguments{
\item{htm}{a string containing a html document.}
}
\value{
a string containing texts converted from the html document.
}
\description{
Wipe out tags in a html document and return a text in it
}
\examples{
text = htm2txt("<html><body>html texts</body></html>")
text = htm2txt(paste(c("<p>html</p>", "texts")))
}
